refactor FileSystem and Implementation into /persistence (#222)
