+++ app/src/main/java/com/github/fribourgsdp/radio/data/User.kt
@@ -2,8 +2,7 @@ package com.github.fribourgsdp.radio.data
 
 
 import android.content.Context
-import com.github.fribourgsdp.radio.AppSpecificFileSystem
-import com.github.fribourgsdp.radio.FileSystem
+import com.github.fribourgsdp.radio.persistence.*
 import com.github.fribourgsdp.radio.FirestoreDatabase
 import com.github.fribourgsdp.radio.util.SetUtility
