+++ app/src/main/java/com/github/fribourgsdp/radio/persistence/AppSpecificFileSystem.kt
@@ -1,4 +1,4 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.persistence
 
 import android.content.Context
 import java.io.File
